:zap: Optimize package size: Avoid bundling chrome-driver and chrome-headless-shell

+++ composeApp/src/commonMain/kotlin/com/crosspaste/utils/RetryUtils.kt
@@ -5,6 +5,6 @@ expect fun getRetryUtils(): RetryUtils
 interface RetryUtils {
     fun <T> retry(
         maxRetries: Int,
-        block: () -> T,
+        block: (Int) -> T,
     ): T?
 }
